+++ packages/transaction-pay-controller/src/strategy/relay/constants.ts
@@ -9,6 +9,7 @@ export const RELAY_EXECUTE_URL = `${RELAY_URL_BASE}/execute`;
 export const RELAY_QUOTE_URL = `${RELAY_URL_BASE}/quote`;
 export const RELAY_STATUS_URL = `${RELAY_URL_BASE}/intents/status/v3`;
 export const HYPERLIQUID_EXCHANGE_URL = 'https://api.hyperliquid.xyz/exchange';
+export const HYPERLIQUID_INFO_URL = 'https://api.hyperliquid.xyz/info';
 export const RELAY_POLLING_INTERVAL = 1000; // 1 Second
 export const TOKEN_TRANSFER_FOUR_BYTE = '0xa9059cbb';